requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pycryptodome>=3.20.0
python-multipart>=0.0.6
aiofiles>=23.0
websockets>=12.0